Moved Team.js and Team.test.js into dir Team
